+++ core/src/main/kotlin/io/specmatic/core/Substitution.kt
@@ -6,7 +6,9 @@ import io.specmatic.core.value.Value
 
 interface Substitution {
     fun isDropDirective(value: Value): Boolean
+    fun resolveIfLookup(value: Value): Value
+    fun substitute(value: Value): ReturnValue<Value>
     fun resolveIfLookup(value: Value, pattern: Pattern): Value
     fun substitute(value: Value, pattern: Pattern, key: String?): ReturnValue<Value>
-    fun upsertStoreUsing(originalValue: Value, runningValue: Value): Substitution = this
+    fun upsertStoreUsing(originalValue: Value, runningValue: Value): Substitution
 }